obs-outputs: Add cast to fix build with Clang 19
Clang 19 has become more strict about mixing different enum types, which
resulted in an error building multimedia/obs-studio on FreeBSD:

/wrkdirs/usr/ports/multimedia/obs-studio/work/obs-studio-30.2.3/plugins
    /obs-outputs/flv-mux.c:659:37: error: bitwise operation between
    different enumeration types ('enum multitrack_type_t' and 'enum
    packet_type_t') [-Werror,-Wenum-enum-conversion]

+++ plugins/obs-outputs/flv-mux.c
@@ -626,7 +626,7 @@ void flv_packet_metadata(enum video_id_t codec_id, uint8_t **output, size_t *siz
 	 * The default trackId is 0.
 	 */
 	if (is_multitrack) {
-		s_w8(&s, MULTITRACKTYPE_ONE_TRACK | PACKETTYPE_METADATA);
+		s_w8(&s, (uint8_t)MULTITRACKTYPE_ONE_TRACK | (uint8_t)PACKETTYPE_METADATA);
 		s_w4cc(&s, codec_id);
 		// trackId
 		s_w8(&s, (uint8_t)idx);
